+++ google2/src/test/scala/org/broadinstitute/dsde/workbench/google2/mock/FakeGoogleStorageInterpreter.scala
@@ -15,7 +15,7 @@ import com.google.cloud.storage.Storage.{
   BucketSourceOption,
   BucketTargetOption
 }
-import com.google.cloud.storage.{Acl, Blob, BlobId, BucketInfo, Cors, Storage, StorageClass}
+import com.google.cloud.storage.{Acl, Blob, BlobId, BucketInfo, Cors, NotificationInfo, Storage, StorageClass}
 import com.google.cloud.{Identity, Policy}
 import fs2.{Pipe, Stream}
 import org.broadinstitute.dsde.workbench.RetryConfig
@@ -240,6 +240,12 @@ class BaseFakeGoogleStorage extends GoogleStorageService[IO] {
                                   bucketSourceOptions: List[BucketSourceOption]
   ): Stream[IO, List[IamPermission]] =
     localStorage.testIamPermissions(bucketName, permissions, traceId, retryConfig, bucketSourceOptions)
+
+  override def createNotificationIfNotExists(bucketName: GcsBucketName,
+                                             notification: NotificationInfo,
+                                             traceId: Option[TraceId],
+                                             retryConfig: RetryConfig
+  ): Stream[IO, Unit] = Stream.empty
 }
 
 object FakeGoogleStorageInterpreter extends BaseFakeGoogleStorage